Add `Code/export_data.R`
First pass at a script that loads everything from the databases, selects
a subset of subjects, and saves to a timestamped `.csv` file. The idea
is that that then becomes a reproducible reference point, rather than
the frequently updated database.
  Note that as yet the inclusion/exclusion do not match the analysis
plan, as it's not decided exactly where is best to apply these.