Remove short/full session confound variable
This seems to have been dealt with by excluding the screening session.
In short, the worry was that `global_z` was based on different numbers
of tests, and we may want to account for that (see e.g.
<https://github.com/nzbri/chchpd/issues/14>). However, the short
sessions do not have these measures at all, so, given that the screening
sessions are already excluded, the `full_assessment` variable does not
add anything:
```
> full_data %>% filter(full_assessment == FALSE) %>% count(global_z)
  global_z     n
     <dbl> <int>
1       NA   289
```